scripts: sign_rproc_fw: rework structure to add tlv blob
Instead of defining specific blob for the hash , public key and
adding property in header, create a TLVs listing properties.
These elements are integrated in a new "trailer" chunk.
This trailer is added for the signature computing.

The platform TLV is just defined but empty. This will be added in
a separate commit

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: Id3a38b66e276ce03b9a4f89d08f1e459a9d1c04b
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/265646
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>