plat-stm32mp2: add external device tree support
Adds the possibility to use and external repository for device tree source
files for this platform. By default, this external path is set to
external-dt/optee.

If a device tree is not found in OP-TEE source code, search it in the
external repository, if any.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I61c0b0e85e2b7431e928b35e24836e8555566e9a
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/304784
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Domain-Review: Lionel DEBIEVE <lionel.debieve@foss.st.com>