drivers: stm32_iwdg: drop call of stm32mp_gic_set_end_of_interrupt()
In the early interrupt handler we run panic(), which does not
return.
There is no need to acknowledge the interrupt in GIC.
If and when in future the panic() would be replaced by other
action, it would be up to the new code to handle the GIC, if
needed.

Drop the call to stm32mp_gic_set_end_of_interrupt().

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: Iebd1cb99ff16e7a3a807ca0ece0b26a8d36da951
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/313638
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Domain-Review: Gatien CHEVALLIER <gatien.chevallier@st.com>