drivers: stm32_iwdg: rework the handling of IWDG_FLAGS_ENABLED
The watchdog is either already started (by OTP HW or FSBL) or started
by this driver. There is no reason to read again the registers to
detect if the watchdog is enabled.

Read only once the register at probe and keep up-to-date the flag
IWDG_FLAGS_ENABLED during the life time of the driver.

While there,
- rename as iwdg_wdt_is_enabled() the function is_enable() and
  properly compute the returned bool value;
- add the function iwdg_wdt_set_is_enabled() to set the flag
  IWDG_FLAGS_ENABLED.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I7eaa6fd506baa3acf49de96c8e10336230fa05d9
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/313642
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Domain-Review: Gatien CHEVALLIER <gatien.chevallier@st.com>