clk: stm32mp: add support of pinctrl and clock for RCC node
Add a generic stm32mp clock driver to support the OP-TEE initialization of
pinctrl and clock defined in the RCC sub-node with compatible "st,rcc".

This driver is used to configure in the board device tree the required
RCC resources as:
- pincontrol to manage Micro-controller clock output (MCO) pins,
  for example to provide this MCO clock on external device
- critical clocks, not managed by other drivers in OP-TEE

This sub-node is optional and avoids to manage the OP-TEE specific
compatibles "st,stm32mp13-rcc-mco" and "st,stm32mp15-rcc-mco",
with a more acceptable binding.

The configuration for this sub node is only done when the early DT drivers
are available, the clock provider (SoC RCC driver) and the pin-controller.

I also align the binding file with kernel one.

Signed-off-by: Gabriel Fernandez <gabriel.fernandez@foss.st.com>
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I48cf65d157be7a3b05cac3e3375c581d2bbcfebc
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/312741
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>