module: optee: clock: reserved clock return stubbed duty cycle
Changes optee clock module to return a dummy 50% duty cycle value to
prevent an error trace in SCP-firmware when client retrieves the duty
cycle of a reserved clock (a defined SCMI clock instance that client
is not expected to manipulated).

This is needed because Linux kernel queries duty cycle values for all
clocks even those not consumed as reserved SCMI clocks are.

Change-Id: Ifd53b80783648df40516e4e1d997e771438a6899
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/SCP-firmware/+/337948
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/377933
Tested-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Domain-Review: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>