plat-stm32mp2: add a memory region reconfiguration API
Add stm32_rif_reconfigure_mem_region() API that can be used to
reconfigure a memory region whether it is under RISAB or RISAF logic.
If not configuration is specified, the access is granted for OP-TEE.
If a configuration is specified, it is applied and the configuration
pointer is freed.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I537836f14fa2aa00ec5e004c58895566441f2027
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/356790
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Tested-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>