stm32mp25: send a pulse on NRSTC1MS pin for DBG_STANDBY
A debugger can program DBGMCU_CR.DBG_STANDBY to keep enabled
power and clocks during standby and to stay connected.

But in this case the PWR lines, as PWR_CPU_ON and PWR_ON, are not managed
and the ROM code hangs at standby wake-up because the regulators used
by the boot devices are not available; normally they are re-enabled by
STPMIC25  based on its pwrctl lines connected to PWR_ON.

So OP-TEE need to send a reset pulse on NRSTC1MS pin at suspend as
workaround of to this issue; this pulse guarantees that the power
sequence requested by the software is performed by STPMIC25 even if the
low power modes is simulated in PWR and the sequence not performed by the
hardware.

The pulse duration is configured in suspend of RCC clock driver, with
C1MSRD in RCC_C1MSRDCR register and default value 0x0: No pulse.

This patch also add a message once to inform the simulated low power mode
because this simulated mode requested for debugger connection can cause or
mask issues.

Change-Id: I91081add72dde14d2466faa04587f5c55c94b9ad
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/364374
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>